A LOT of changes for the islands
Camera Controller- follows the character
Dialogue - a pop up box with text that is shown only when near the lizard or any creature from current island
Idle Animation - unfortunately with the time crunch i was only able to create an idle animation that plays on loop
Design- Mailman and Island design finally implemented!
Rigging Mailman - added a bone tool Base bone connects to legs; Body bone connects to arms;(this is how the animation is controlled).
Loading Manager - a little delay to have a nice animation when going through the scenes
Everything made into Prefabs - so I can make changes and have it updated everywhere without needing to go back to the code
Copy and pasted everything to the other islands- now i just need to change the dialogue and design if the time calls for it